backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.1
anthropic==0.25.1
supabase==2.4.3
apscheduler==3.10.4
pydantic==2.7.1
python-dotenv==1.0.1
